+++ packages/terra-clinical-site/CHANGELOG.md
@@ -8,6 +8,7 @@ Unreleased
 ----------
 ### Added
 * Add component names to Props Table titles
+* Update to LabelValueView examples
 
 2.0.0 - (February 19, 2018)
 ----------